Move modules to lib folder
