+++ src/agent/index.js
@@ -1,29 +1,29 @@
 /* eslint-disable comma-dangle */
 'use strict';
 
-const stalker = require('./debug/stalker');
-const android = require('./java/android');
-const classes = require('./info/classes');
+const android = require('./lib/java/android');
+const classes = require('./lib/info/classes');
 const config = require('./config');
-const darwin = require('./darwin');
-const debug = require('./debug');
-const disasm = require('./disasm');
-const expr = require('./expr');
-const fs = require('./fs');
-const info = require('./info');
-const io = require('./io');
-const interceptor = require('./debug/interceptor');
-const java = require('./java');
-const log = require('./log');
-const lookup = require('./lookup');
-const memory = require('./memory');
+const darwin = require('./lib/darwin');
+const debug = require('./lib/debug');
+const disasm = require('./lib/disasm');
 const dump = require('./dump');
-const r2 = require('./r2');
-const search = require('./search');
-const sys = require('./sys');
-const swift = require('./darwin/swift');
-const trace = require('./debug/trace');
-const utils = require('./utils');
+const expr = require('./expr');
+const fs = require('./lib/fs');
+const info = require('./lib/info');
+const io = require('./io');
+const interceptor = require('./lib/debug/interceptor');
+const java = require('./lib/java');
+const log = require('./log');
+const lookup = require('./lib/info/lookup');
+const memory = require('./lib/debug/memory');
+const r2 = require('./lib/r2');
+const search = require('./lib/search');
+const stalker = require('./lib/debug/stalker');
+const sys = require('./lib/sys');
+const swift = require('./lib/darwin/swift');
+const trace = require('./lib/debug/trace');
+const utils = require('./lib/utils');
 
 // r2->io->frida->r2pipe->r2
 let _r2 = null;
