+++ src/agent/lib/debug/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const r2 = require('./r2');
-const sys = require('./sys');
-const utils = require('./utils');
+const r2 = require('../r2');
+const sys = require('../sys');
+const utils = require('../utils');
 
 /* Globals */
 const newBreakpoints = new Map();
