+++ src/agent/lib/debug/memory.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const expr = require('./expr');
-const globals = require('./globals');
-const utils = require('./utils');
+const expr = require('../expr');
+const globals = require('../globals');
+const utils = require('../utils');
 
 function listMemoryRanges () {
   return listMemoryRangesJson()
