+++ src/agent/lib/disasm.js
