+++ src/agent/lib/dump.js
