+++ src/agent/lib/expr.js
