+++ src/agent/lib/fs.js
