+++ src/agent/lib/globals.js
