+++ src/agent/lib/info/lookup.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const utils = require('./utils');
+const utils = require('../utils');
 
 function lookupDebugInfo (args) {
   const o = DebugSymbol.fromAddress(ptr('' + args));
