+++ src/agent/lib/strings.js
